Adding two integers may result in loss of precision

+++ systematic-trading-strategy/src/main/java/com/systematic/trading/strategy/confirmation/TradingStrategyConfirmedBy.java
@@ -82,6 +82,6 @@ public class TradingStrategyConfirmedBy implements Confirmation {
 	}
 
 	private LocalDate latestConfirmationDate( final DatedSignal anchor ) {
-		return anchor.getDate().plusDays(delayUntilConfirmationRange + confirmationDayRange);
+		return anchor.getDate().plusDays((long) delayUntilConfirmationRange + confirmationDayRange);
 	}
 }
